Fixes stub build break and maybe also arm (#31)
